pyproject.toml
[tool.poetry]
name = "techfin"
# Version is decided automatically by `poetry dynamic-versioning`
version = "0.0.0+124.7eb61a1"
description = ""
authors = ["Inekto <contact@inekto.fr>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sentry-sdk = "^1.44.1"
loguru = "^0.7.2"
prometheus-fastapi-instrumentator = "^7.0.0"
prometheus-client = "^0.20.0"
factory-boy = "^3.3.0"
sqlmodel = "^0.0.16"


[tool.poetry.group.zenioo.dependencies]
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
requests = "^2.31.0"
python-decouple = "^3.8"


[tool.poetry.group.backend.dependencies]
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
requests = "^2.31.0"
python-decouple = "^3.8"
httpx = "^0.27.0"
firebase-admin = "^6.5.0"
hubspot-api-client = "^9.0.0"


[tool.poetry.group.assurland.dependencies]
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
requests = "^2.31.0"
python-decouple = "^3.8"
httpx = "^0.27.0"
pydantic = "^2.6.4"
email-validator = "^2.1.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.5"
pre-commit = "^3.7.0"
locust = "^2.24.1"


[tool.poetry.group.release.dependencies]
versioneer = "^0.29"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry-dynamic-versioning]
enable = true
vcs = 'git'
format-jinja = "{% if distance == 0 %}{{ base }}{% else %}{{ base }}+{{ distance }}.{{ commit }}{% endif %}"

[tool.poetry-dynamic-versioning.substitution]
files = ["./**/__version__.py"]
